requirements.txt
Django==1.10.1
django-polymorphic==1.0.1
factory-boy==2.6.1
google-api-python-client==1.4.2
mock==1.3.0
oauth2client==1.5.2
